refactor: extract CSS class decorator from service
